sletter gammel søknadservice og tilhørende tester
